requirements.txt
psutil>=5.9.0
watchdog>=2.1.9
scapy>=2.5.0
pywin32>=305
pefile>=2023.2.7
matplotlib==3.5.1
Pillow==9.0.1
pandas>=2.0.0
scikit-learn>=1.2.2
joblib>=1.2.0
tkinter
numpy>=1.24.0